requirements.txt
streamlit>=1.18.0
pdfplumber>=0.5.28
PyPDF2>=3.0.0
pandas>=1.4.0
python-docx>=0.8.11
openpyxl>=3.0.10
